callbacks for Collation and Traversal node strategies
migrating a refactor which allows collation and traversal strategies to make callbacks using the same logic
